refactor: rm Gitpod-specific README
The main & Gitpod READMEs had become out of sync. It looks like the only
difference was that the Gitpod readme omitted the "Prerequisites" &
"Building this contract" sections. I've added a note to the
"Prerequisites" section for Gitpod users, letting them know they can
skip it. And I think we want to leave in the "Building this contract"
section for everyone.